Adding tests for the new apis to convert between S.Numerics and S.R.Intrinsic vectors (dotnet/coreclr#27481)
* Updating Vector128.AsVector3 to have a functioning implementation

* Adding test template metadata covering the new methods for converting between S.Numerics and S.R.Intrinsics

* Regenerating the existing tests

* Generating the new Vector conversion tests


Commit migrated from https://github.com/dotnet/coreclr/commit/4b902f244e529b7596b1501eea6fdf27753b91ca

+++ src/coreclr/tests/src/JIT/HardwareIntrinsics/General/Vector128_1/As.Byte.cs
@@ -51,43 +51,43 @@ namespace JIT.HardwareIntrinsics.General
             TestLibrary.TestFramework.BeginScenario(nameof(RunBasicScenario));
             Vector128<Byte> value;
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<byte> byteResult = value.AsByte();
             ValidateResult(byteResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<double> doubleResult = value.AsDouble();
             ValidateResult(doubleResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<short> shortResult = value.AsInt16();
             ValidateResult(shortResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<int> intResult = value.AsInt32();
             ValidateResult(intResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<long> longResult = value.AsInt64();
             ValidateResult(longResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<sbyte> sbyteResult = value.AsSByte();
             ValidateResult(sbyteResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<float> floatResult = value.AsSingle();
             ValidateResult(floatResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<ushort> ushortResult = value.AsUInt16();
             ValidateResult(ushortResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<uint> uintResult = value.AsUInt32();
             ValidateResult(uintResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<ulong> ulongResult = value.AsUInt64();
             ValidateResult(ulongResult, value);
         }
@@ -97,43 +97,43 @@ namespace JIT.HardwareIntrinsics.General
             TestLibrary.TestFramework.BeginScenario(nameof(RunGenericScenario));
             Vector128<Byte> value;
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<byte> byteResult = value.As<Byte, byte>();
             ValidateResult(byteResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<double> doubleResult = value.As<Byte, double>();
             ValidateResult(doubleResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<short> shortResult = value.As<Byte, short>();
             ValidateResult(shortResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<int> intResult = value.As<Byte, int>();
             ValidateResult(intResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<long> longResult = value.As<Byte, long>();
             ValidateResult(longResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<sbyte> sbyteResult = value.As<Byte, sbyte>();
             ValidateResult(sbyteResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<float> floatResult = value.As<Byte, float>();
             ValidateResult(floatResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<ushort> ushortResult = value.As<Byte, ushort>();
             ValidateResult(ushortResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<uint> uintResult = value.As<Byte, uint>();
             ValidateResult(uintResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             Vector128<ulong> ulongResult = value.As<Byte, ulong>();
             ValidateResult(ulongResult, value);
         }
@@ -143,70 +143,70 @@ namespace JIT.HardwareIntrinsics.General
             TestLibrary.TestFramework.BeginScenario(nameof(RunReflectionScenario));
             Vector128<Byte> value;
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object byteResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsByte))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<byte>)(byteResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object doubleResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsDouble))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<double>)(doubleResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object shortResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsInt16))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<short>)(shortResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object intResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsInt32))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<int>)(intResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object longResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsInt64))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<long>)(longResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object sbyteResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsSByte))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<sbyte>)(sbyteResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object floatResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsSingle))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<float>)(floatResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object ushortResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsUInt16))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<ushort>)(ushortResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object uintResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsUInt32))
                                     .MakeGenericMethod(typeof(Byte))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<uint>)(uintResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetByte());
+            value = Vector128.Create((byte)TestLibrary.Generator.GetByte());
             object ulongResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsUInt64))
                                     .MakeGenericMethod(typeof(Byte))
