+++ src/coreclr/tests/src/JIT/HardwareIntrinsics/General/Vector128_1/As.Int32.cs
@@ -51,43 +51,43 @@ namespace JIT.HardwareIntrinsics.General
             TestLibrary.TestFramework.BeginScenario(nameof(RunBasicScenario));
             Vector128<Int32> value;
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<byte> byteResult = value.AsByte();
             ValidateResult(byteResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<double> doubleResult = value.AsDouble();
             ValidateResult(doubleResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<short> shortResult = value.AsInt16();
             ValidateResult(shortResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<int> intResult = value.AsInt32();
             ValidateResult(intResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<long> longResult = value.AsInt64();
             ValidateResult(longResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<sbyte> sbyteResult = value.AsSByte();
             ValidateResult(sbyteResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<float> floatResult = value.AsSingle();
             ValidateResult(floatResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<ushort> ushortResult = value.AsUInt16();
             ValidateResult(ushortResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<uint> uintResult = value.AsUInt32();
             ValidateResult(uintResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<ulong> ulongResult = value.AsUInt64();
             ValidateResult(ulongResult, value);
         }
@@ -97,43 +97,43 @@ namespace JIT.HardwareIntrinsics.General
             TestLibrary.TestFramework.BeginScenario(nameof(RunGenericScenario));
             Vector128<Int32> value;
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<byte> byteResult = value.As<Int32, byte>();
             ValidateResult(byteResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<double> doubleResult = value.As<Int32, double>();
             ValidateResult(doubleResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<short> shortResult = value.As<Int32, short>();
             ValidateResult(shortResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<int> intResult = value.As<Int32, int>();
             ValidateResult(intResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<long> longResult = value.As<Int32, long>();
             ValidateResult(longResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<sbyte> sbyteResult = value.As<Int32, sbyte>();
             ValidateResult(sbyteResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<float> floatResult = value.As<Int32, float>();
             ValidateResult(floatResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<ushort> ushortResult = value.As<Int32, ushort>();
             ValidateResult(ushortResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<uint> uintResult = value.As<Int32, uint>();
             ValidateResult(uintResult, value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             Vector128<ulong> ulongResult = value.As<Int32, ulong>();
             ValidateResult(ulongResult, value);
         }
@@ -143,70 +143,70 @@ namespace JIT.HardwareIntrinsics.General
             TestLibrary.TestFramework.BeginScenario(nameof(RunReflectionScenario));
             Vector128<Int32> value;
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object byteResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsByte))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<byte>)(byteResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object doubleResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsDouble))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<double>)(doubleResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object shortResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsInt16))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<short>)(shortResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object intResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsInt32))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<int>)(intResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object longResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsInt64))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<long>)(longResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object sbyteResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsSByte))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<sbyte>)(sbyteResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object floatResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsSingle))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<float>)(floatResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object ushortResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsUInt16))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<ushort>)(ushortResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object uintResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsUInt32))
                                     .MakeGenericMethod(typeof(Int32))
                                     .Invoke(null, new object[] { value });
             ValidateResult((Vector128<uint>)(uintResult), value);
 
-            value = Vector128.Create(TestLibrary.Generator.GetInt32());
+            value = Vector128.Create((int)TestLibrary.Generator.GetInt32());
             object ulongResult = typeof(Vector128)
                                     .GetMethod(nameof(Vector128.AsUInt64))
                                     .MakeGenericMethod(typeof(Int32))
